+++ src/SciChartMemoryDebugWrapper.tsx
@@ -8,14 +8,19 @@ import { SciChartSurface, MemoryUsageHelper } from "scichart";
  * Rendering the component enables Memory Debug Mode and adds extra UI elements to allow easily mount/unmount the child components
  * @remarks Find more info at [Memory Leak Debugging docs](https://www.scichart.com/documentation/js/current/MemoryLeakDebugging.html)
  */
-export function SciChartMemoryDebugWrapper(props: PropsWithChildren<any>): JSX.Element {
+export function SciChartMemoryDebugWrapper(
+    props: PropsWithChildren<{
+        /** Whether the child components are rendered on the first render. Defaults to false */
+        defaultRenderChildren?: boolean;
+    }>
+): JSX.Element {
     useState(() => {
         MemoryUsageHelper.isMemoryUsageDebugEnabled = true;
         SciChartSurface.autoDisposeWasmContext = true;
         SciChartSurface.wasmContextDisposeTimeout = 0;
     });
 
-    const [drawChart, setDrawChart] = useState(false);
+    const [drawChart, setDrawChart] = useState(props.defaultRenderChildren ?? false);
 
     const handleCheckbox: ChangeEventHandler<HTMLInputElement> = e => {
         setDrawChart(e.target.checked);